move DeclarativeTaskTestMixin to TaskTestBase
