+++ tests/python/pants_test/engine/scheduler_test_base.py
@@ -9,14 +9,11 @@ import shutil
 from builtins import object
 
 from pants.base.file_system_project_tree import FileSystemProjectTree
-from pants.build_graph.address import Address
 from pants.engine.nodes import Throw
 from pants.engine.scheduler import Scheduler
 from pants.option.global_options import DEFAULT_EXECUTION_OPTIONS
 from pants.util.contextutil import temporary_file_path
 from pants.util.dirutil import safe_mkdtemp, safe_rmtree
-from pants.util.memo import memoized_method
-from pants.util.meta import classproperty
 from pants_test.engine.util import init_native
 
 
@@ -100,120 +97,3 @@ class SchedulerTestBase(object):
     self.assertTrue(type(resulting_value) is Throw)
 
     raise resulting_value.exc
-
-
-class DeclarativeTaskTestBase(SchedulerTestBase):
-  """???/experimental blah, makes things declarative, whatever"""
-
-  @classproperty
-  def dist_specs(cls):
-    """
-    This is an informally-specified nested dict -- see ../test_ctypes.py for an example. Special
-    keys are 'key' (used to index into `self.target_dict`) and 'filemap' (creates files at the
-    specified relative paths). The rest of the keys are fed into `self.make_target()`. An
-    `OrderedDict` of 2-tuples may be used if targets need to be created in a specific order (e.g. if
-    they have dependencies on each other).
-    """
-    raise NotImplementedError('dist_specs must be implemented!')
-
-  @classproperty
-  def run_before_task_types(cls):
-    """
-    By default, we just use a `BuildLocalPythonDistributions` task. When testing with C/C++ targets,
-    we want to compile and link them as well to get the resulting dist to build, so we add those
-    task types here and execute them beforehand.
-    """
-    return []
-
-  @classproperty
-  def run_after_task_types(cls):
-    """Tasks to run after local dists are built, similar to `run_before_task_types`."""
-    return []
-
-  def populate_target_dict(self):
-    self.target_dict = {}
-
-    # Create a target from each specification and insert it into `self.target_dict`.
-    for target_spec, target_kwargs in self.dist_specs.items():
-      unprocessed_kwargs = target_kwargs.copy()
-
-      target_base = Address.parse(target_spec).spec_path
-
-      # Populate the target's owned files from the specification.
-      filemap = unprocessed_kwargs.pop('filemap', {})
-      for rel_path, content in filemap.items():
-        buildroot_path = os.path.join(target_base, rel_path)
-        self.create_file(buildroot_path, content)
-
-      # Ensure any dependencies exist in the target dict (`dist_specs` must then be an
-      # OrderedDict).
-      # The 'key' is used to access the target in `self.target_dict`.
-      key = unprocessed_kwargs.pop('key')
-      dep_targets = []
-      for dep_spec in unprocessed_kwargs.pop('dependencies', []):
-        existing_tgt_key = self.dist_specs[dep_spec]['key']
-        dep_targets.append(self.target_dict[existing_tgt_key])
-
-      # Register the generated target.
-      generated_target = self.make_target(
-        spec=target_spec, dependencies=dep_targets, **unprocessed_kwargs)
-      self.target_dict[key] = generated_target
-
-  @memoized_method
-  def _synthesize_task_types(self, task_types=()):
-    return [
-      self.synthesize_task_subtype(tsk, '__tmp_{}'.format(tsk.__name__))
-      # TODO: make @memoized_method convert lists to tuples for hashing!
-      for tsk in task_types
-    ]
-
-  def _all_specified_targets(self):
-    return list(self.target_dict.values())
-
-  def _scheduling_context(self, **kwargs):
-    scheduler = self.mk_scheduler(rules=self.rules())
-    return self.context(scheduler=scheduler, **kwargs)
-
-  def _retrieve_single_product_at_target_base(self, product_mapping, target):
-    product = product_mapping.get(target)
-    base_dirs = list(product.keys())
-    self.assertEqual(1, len(base_dirs))
-    single_base_dir = base_dirs[0]
-    all_products = product[single_base_dir]
-    self.assertEqual(1, len(all_products))
-    single_product = all_products[0]
-    return single_product
-
-  def _create_task(self, task_type, context):
-    return task_type(context, self.test_workdir)
-
-  def invoke_tasks(self, **context_kwargs):
-    run_before_synthesized_task_types = self._synthesize_task_types(tuple(self.run_before_task_types))
-    run_after_synthesized_task_types = self._synthesize_task_types(tuple(self.run_after_task_types))
-    all_synthesized_task_types = run_before_synthesized_task_types + [
-      self._testing_task_type,
-    ] + run_after_synthesized_task_types
-
-    context = self._scheduling_context(
-      for_task_types=all_synthesized_task_types,
-      **context_kwargs)
-    self.assertEqual(set(self._all_specified_targets()), set(context.build_graph.targets()))
-
-    run_before_task_instances = [
-      self._create_task(task_type, context)
-      for task_type in run_before_synthesized_task_types
-    ]
-    current_task_instance = self._create_task(
-      self._testing_task_type, context)
-    run_after_task_instances = [
-      self._create_task(task_type, context)
-      for task_type in run_after_synthesized_task_types
-    ]
-    all_task_instances = run_before_task_instances + [
-      current_task_instance
-    ] + run_after_task_instances
-
-    for tsk in all_task_instances:
-      tsk.execute()
-
-    return (context, run_before_task_instances, current_task_instance, run_after_task_instances)
